Tests/LibWeb/Text/expected/HTML/DataTransfer-clearData-stale-item.txt

items.length after clearData = 0
stale item kind = ""
stale item type = ""
stale item getAsFile() = null
